plan_modifiers: move PreserveEmptyStringAsNull into plan_modifiers.go
Updates #cleanup
